+++ CHANGE_LOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - source-jars plug-in which provides tasks to build source and javadoc jars and adds them to the project's archives artifact configuration
 - Integrate source-jars with maven-publish plug-in as published artifacts
 - metadata-base plug-in which provides a DSL for re-usable project meta data values
+- Add metadata-base plug-in to multi-module-library plug-in setup for all projects
+- metadata-pom plug-in which applies meta data in metadata-base DSL to generated Maven POM files
+- Add metadata-pom plug-in to multi-module-library plug-in setup for all projects
 
 ## [0.1.0]
 ### Added